change: stream-typed cosocket pool supports aborting connect() when exceeding the pool size limit
added new nginx config directive "lua_socket_pool_exceeding" to
control the behavior when the stream-typed cosocket pool exceeds its
size limit. right now two values are accepted: "ignore" and "abort". For
the former, connect() will proceed to establish a new connection and for
the latter, connect() will return the error "exceeding pool size"
immediately. The former is the default for now.

The pool size limit is no longer specified by the second argument of the
"setkeepalive" method, but rather as the "pool_size" option in the
(optional) option table in the "connect" method, for instance,

    local ok, err = sock:connect("127.0.0.1", 6379, { pool_size = 10 })

The second argument of the "setkeepalive" method is now always ignored.