Discard request-body if it's not read at all.
This is necessary in the case of pipelined request; the un-discarded request-body would otherwise linger to the next request, and will be mistakenly recognized as part of the header of the next pipelined request.

This change is to explicitly discard request-body for two directives: access_by_lua and content_by_lua. It leave out two situations without fixed: when the directive handler returns NGX_DONE and NGX_EAGAIN. Quite honestly, I don't know for sure how to fix right at this moment. But I think I catch the most common cases.