feat: ngx.log accepts table with __tostring metafield
Extend ngx.log() to allow table arguments when those have a '__tostring'
metafield. If they do not, ngx.log will raise an error (similar to
previous behavior when a table was given to ngx.log).

When a table with '__tostring' metafield is given, the metamethod will
be called through luaL_callmeta().