Skip setting multi-value headers for bad requests
When setting a multi-value header during bad requests, the value isn't
added into the header's array. Despite failing to be set, the size of the
array is still incremented. Later, if Nginx attempts to iterate over the
array, a segfault will occur.

e21d9b5 established a pattern of silently ignoring the header being set.
I've moved the check for a bad request into `ngx_http_lua_set_input_header`
so that setting a header value is ignored for both types of headers.