Revert "bugfix: ngx.escape_uri: we did not escape URI reserved chars. #1124"
Commit f170505186ff61af36b3e126772b671793af9428 breaks the compatibility
with RFC 3986. Here is two reasons:
1. Quote from RFC 3986 Section 2.2:
> A subset of the reserved characters (gen-delims) is used as
delimiters of the generic URI components described in Section 3

Note that RFC 3986 says 'a subset of the reserved characters
(gen-delims)', not all the reserved characters. The characters escaped
in that commit are 'sub-delims'. They are not required to be escaped
according to Section 2.2.

2. Refer to RFC 3986 "Appendix A.  Collected ABNF for URI", sub-delims
could be used as part of query and other components. This use case shows
that sub-delims are valid in some component of URI. Therefore, it would
be better if we don't escape them for URI component.

Signed-off-by: Yichun Zhang (agentzh) <agentzh@gmail.com>