bugfix: old location's code would still be run after filter finalization
When calling output API functions like ngx.print with
If-Unmodified-Since(or If-Match) in the request headers,
the ngx_http_not_modified_filter_module will test the "If-Unmodified-Since",
if failed, ngx_http_filter_finalize_request will be called and the contexts of
all http modules will be cleared, after that, if we call ngx.flush(), we
will get the error "no request ctx found". Worse more, when error_page
was configurated, the old location's code still would be run.

Please see the github issue
https://github.com/openresty/lua-nginx-module/issues/1131
for the details.