change: switched to SSL_version() calls from TLS1_get_version().
TLS1_get_version is a simple wrapper for SSL_version that returns 0
when used with DTLS. However, it was removed from BoringSSL in 2015
so instead use SSL_version directly. Note: BoringSSL is never an
officially supported target for this module.

ngx_http_lua_ffi_ssl_get_tls1_version can never be reached with DTLS
so the behaviour is the same.

Signed-off-by: Yichun Zhang (agentzh) <agentzh@gmail.com>