bugfix: the lightuserdata mask was broken in WIN64.
In WIN64, the pointer type is 64 bits but the long type is 32 bits.
Since WIN64 only supports x64 and IA64, we could confirm that it doesn't
need the mask.

If we need the mask in the future, we might need to use a C compiler
which supports C99 (like MSVC 2013+ or any other modern C compiler) to build
the WIN64 version.