Generalise shared-dict argument error message
Have reworded the error message for the lua_shared_dict directive to
ensure there is no confusion when there are more than 2 arguments passed
into it. The confusion here was around the use of the word only which
suggests you have received less arguments than required amount, however
when passed 3 or more this causes the error message to become confusing.
```
failed to run set_by_lua*: set_by_lua:3: expecting exactly two arguments, but only seen 3
```
Have replaced, 'only seen' with 'received' to ensure that the error
message is more generic and doesnt lead to confusion when reading the message.