change: we now avoid generating the Content-Type response header when getting all response headers.
The previous behavior assumed that nginx will always generate the
Content-Type header if it is missing. However in some cases (such as the
upstream response not having a Content-Type header), nginx will not
generate the Content-Type header itself.

Thus, generating the Content-Type when getting response headers
may create an unexpected Content-Type header in the response.

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>