bugfix: avoided running init_by_lua* when test Nginx configuration which contains lua_shared_dict directive.
In commit 7c2b58e, we avoided running init_by_lua* in signaller processes
and when testing the Nginx configuration. However, when the Nginx
configuration contains lua_shared_dict directive, the execution of
init_by_lua* will be postponded when testing the configuration.
This case was not handled and led to the segmentation fault reported in
GitHub issue #1462.