bugfix: fixed a segfault introduced in 7c2b58e when testing an nginx configuration containing 'lua_shared_dict' directive(s).
In commit 7c2b58e, we avoided running `init_by_lua*` in signaller processes and
when testing the Nginx configuration. However, when the Nginx configuration
contains `lua_shared_dict` directives, the execution of `init_by_lua*` will be
postponded when testing the configuration.

This case was not handled and led to the segmentation fault reported in #1462.