optimize: avoided using strings as code cache keys.
This optimization takes advantage of the `luaL_ref()` C API to generate
unique references for Lua buffers (inline and file). This avoids runtime
calls to `lj_str_new()` while invoking Lua handler.

By doing so, we observe reduced CPU utilization and increased throughput
(~2% to ~5% in preliminary tests) in a simplistic hello world handler.
Other benefits of this patch include the simplification of the code
cache keys generation and reduced memory consumption.

In this implementation, a `LUA_REFNIL` code cache reference means that the
associated chunk has not been cached yet. A `LUA_NOREF` code cache
reference applies to Lua file paths and means that this particular file
path has nginx variables and should thus generate its own cache key
(identical to the current implementation).

This patch also fixes an isue with `set_by_lua_file` handlers if those
were specified with nginx variables. We added a new test case for this
behavior which fails on the master branch of this repository (v0.10.14),
but passes with this patch.