bugfix: skipped 'ssl_session_store_by_lua*' and 'ssl_session_fetch_by_lua*' when using TLS 1.3.
Previously, we used the OpenSSL 1.1.1 ClientHello callback to do ssl
session fetching non-blockingly. However, this way cannot handle an edge
case: the ssl session resumption via session ticket might fail, and the
client fallbacks to session ID resumption. The ClientHello callback is
run too early to know if the client will fallback to use session ID
resumption.

Therefore, we have to take back the OpenSSL sess_set_get_cb_yield patch
and upgrade it to adapt OpenSSL 1.1.1, which was done in our
openresty/openresty repository.

Now, this means that for the time being, we must skip
`ssl_session_(fetch|store)_by_lua*` for TLS 1.3. It is possible to
support PSK with session ID in TLS 1.3., but we need to modify a number
of functions to pass the result up, which will make the patch too
complex to maintain. Since PSK with session ticket is supported,
supporting PSK with session ID is not so profitable. If someone needs
this feature, they can contribute it themselves.

Thanks Yongjian Xu and crasyangel for their help.

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>