bugfix: when recreating balancer buffer, remove existing buffers first if they exist.
This bug was discovered first by Kong users in
https://github.com/Kong/kong/issues/6212. After investigation, it seems
that the cause is that when calling `create_request` function
multiple times, it may incorrectly consider the buffer created in the
last call request body data, causing both the old and new request buffer
to be sent to the upstream.

This PR checks the request buffer chain link before calling the
`create_request` function. If it contains a buffer generated previously,
that buffer is skipped first so the new buffer that will be generated
falls in at the correct location.