Set headers in the tail of a list rather than in the head
It seems that ngx_http_variable_headers_internal function in
nginx/src/http/ngx_http_variables.c relies on the fact that the
last header values is stored in the last element of a list.
Therefore we should avoid changing a random element in the head of
a list of headers and either change the last element or add a new
element to the list.