bugfix: respect max retry after using balancer pool
In the balancer phase, when obtaining a connection from the upstream
connection pool, the `cached` attribute of the peer connection is set
to 1(`pc->cached = 1;`), indicating that the connection is obtained
from the cache.

If an error occurs during the use of this connection, such as
"upstream prematurely closed connection" the system will increase the
`tries` attribute of the peer connection by executing
`u->peer.tries++`.

This PR restores tries by callbacks to the balancer when
`u->peer.tries++` is unexpectedly set.

Signed-off-by: tzssangglass <tzssangglass@gmail.com>