ngx_http_lua_ffi_ssl_get_client_hello_ciphers()
Partially inspired by:
	https://github.com/naofumi0628/haproxy/blob/fefb9e37714bd2e3ad2adc3a321e165fc1dafae2/src/ssl_sock.c#L2252

Relevant:
	https://github.com/fooinha/nginx-ssl-ja3/issues/64
        https://github.com/openssl/openssl/issues/27580

And especially:
	https://github.com/openresty/lua-nginx-module#:~:text=after%20SSL%20handshake%2C-,the%20ngx.ctx%20created,-in%20ssl_certificate_by_lua*

It might be pointless for me to pull all this data into Lua-land if I don't find a way to store those values.
I need some kind of ngx.ctx but related not a request but to a ngx_ssl_connection_t, instead of a request.